Simulating upcoming request insertion

+++ src/ProxyBwapi/UnitClasses/UnitClassProxy.scala
@@ -66,7 +66,7 @@ abstract class UnitClassProxy(val bwapiType: UnitType) {
   val producesLarva            = bwapiType.producesLarva
   val regeneratesHP            = bwapiType.regeneratesHP
   lazy val requiredTechRaw     = Techs.get(bwapiType.requiredTech)
-  lazy val requiredUnits       = bwapiType.requiredUnits.asScala.map(pair => (UnitClasses.get(pair._1), pair._2))
+  lazy val requiredUnits       = bwapiType.requiredUnits.asScala.map(pair => (UnitClasses.get(pair._1), pair._2.toInt))
   val requiresCreep            = bwapiType.requiresCreep
   val requiresPsi              = bwapiType.requiresPsi
   lazy val techsWhat           = bwapiType.researchesWhat.asScala.map(Techs.get)
@@ -83,7 +83,7 @@ abstract class UnitClassProxy(val bwapiType: UnitType) {
   val turnRadius               = bwapiType.turnRadius
   lazy val upgrades            = bwapiType.upgrades.asScala.map(Upgrades.get)
   lazy val upgradesWhat        = bwapiType.upgradesWhat.asScala.map(Upgrades.get)
-  lazy val whatBuilds          = new Pair(UnitClasses.get(bwapiType.whatBuilds.getKey), bwapiType.whatBuilds.getValue)
+  lazy val whatBuilds          = new Pair(UnitClasses.get(bwapiType.whatBuilds.getKey), bwapiType.whatBuilds.getValue.toInt)
   val width                    = bwapiType.width
   val race                     = bwapiType.getRace
   val airWeaponRaw             = bwapiType.airWeapon
